+++ src/main/java/dev/ngspace/hudder/hudderv3/instructions/compiler/StringInstruction.java
@@ -3,8 +3,8 @@ package dev.ngspace.hudder.hudderv3.instructions.compiler;
 import org.objectweb.asm.Label;
 
 import dev.ngspace.hudder.exceptions.CompileException;
-import dev.ngspace.hudder.hudderv3.V3ExecuteMethodWriter;
 import dev.ngspace.hudder.hudderv3.asm.V3ClassWriter;
+import dev.ngspace.hudder.hudderv3.asm.V3ExecuteMethodWriter;
 
 public class StringInstruction extends Instruction {
